Update Helm to v3.7.2
This commit updates Helm to 3.7.2, in an attempt to get to a v3.7.x
release range _without_ any memory issues (see #345), which should have
been addressed in this release.

The change in replacements has been cross-checked with the dependencies
of Helm (and more specifically, the Oras project), and confirmed to not
trigger any warnings using `trivy`.

Signed-off-by: Hidde Beydals <hello@hidde.co>